Add files via upload
This was a comprehensive lab project from my Introduction to Computer Science 1 course at the University of Texas at El Paso. It was composed of 3 challenges, the 1st being to create a program that runs the Luhn Method. The second challenge was to read from a file containing 10 credit card numbers and print to the screen whether they are valid or not. The program then writes to a new file the names and credit card numbers of the individuals with invalid cards. The last challenge was a bonus to be able to produce a new valid credit card number that passes the Luhn Method every time the program is run.